Update declared code dependencies
I've started by removing the "requirements.txt" file in favor of a
Pipfile (at the end of the day, the README requires the use of PipEnv so
that would have been created). I've then removed the fixed versions and
installed the latest versions compatible with Python 3.9. Same goes for
Node.js dependencies. A code update is now required.

I've also tried to pin compatible versions of Rasa and Socketio, and
I've replaced the MySQL connector in use with the official one.