Remove manipulation of paths as strings
Originally, paths were treated as simple strings that were manually
joined by obtaining the platform-specific path separator and manually
appending it to strings. This may lead to problems in the future (e.g.,
if a path separator is appended to a variable that already contains a
terminal path separator).

To fix this, using os.path.join would have sufficed. I've opted,
instead, for using the standard module pathlib, to make the code more
readable and to allow for further manipulation of paths if needed.