Introduce connection string builder
Introduced a simple builder class for creating connection strings. This
allows to ensure consistent formatting of the connection string, and
allows also for a leaner code as the "conditional" adding of components
to the string is hidden inside a class. Finally, this also ensure that
when adding a component to the string, the component is terminated with
a semicolon.

Known issue: multiple calls to the same method will add the same
component multiple times. This needs to be fixed, but it's fine for now.