Allow the bot to manage multiple diagrams
Previously, requests from a user could overwrite requests from another
user if they were concurrent. To avoid this issue, we now put a
timestamp to each generated image, and we also add the session id to the
filename to ensure uniqueness.

Furthermore, while I was at it, I added a little "if" in which we can
easily add more command-based answers (like the one used for the pie
chart, which is now mapped to the command /pie-chart <file>).